refactor(components): refactor directory systems
