--- src/components/Footer.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export function Footer() {
-  return <div className="footer">Copyright. 2019 Shinyaigeek</div>;
-}
