+++ src/front/Home/components/WelcomePage/WelcomePage.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+
+import { ShinyaigeekAnimation } from "../../../components/ShinyaigeekAnimation/ShinyaigeekAnimation";
+import { ScrollDown } from "../ScrollDown/ScrollDown";
+import { StarStream } from "../StartStream/StarStream";
+
+export function WelcomePage() {
+  return (
+    <div className="welcome">
+      <img className="banana" src="https://storage.cloud.google.com/blog_assets_shinyaigeek/static/banana.png" alt="banana" />
+      <ScrollDown />
+      <ShinyaigeekAnimation />
+      <StarStream />
+    </div>
+  );
+}
