+++ src/front/components/DecorationTag/DecorationTag.tsx
