+++ src/front/components/Divider/Divider.tsx
