+++ src/front/components/Footer/Footer.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export function Footer() {
+  return <div className="footer">Copyright. 2020 Shinyaigeek</div>;
+}
